Combine record and cast modules
